Перенес логику создания вопросов в новый класс Создал файлы для структур QuizStepViewModel,  QuizResultsViewModel, QuizQuestion и перенесл структуры в эти файлы
