+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+// для состояния "Результат квиза"
+struct QuizResultsViewModel {
+  // строка с заголовком алерта
+  let title: String
+  // строка с текстом о количестве набранных очков
+  let text: String
+  // текст для кнопки алерта
+  let buttonText: String
+}
